django_test/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
bleach==6.0.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.4
django-constance==3.1.0
django-picklefield==3.1
django-summernote==0.8.20.0
django-treasuremap==0.3.4
djangorestframework==3.14.0
et-xmlfile==1.1.0
geojson==2.5.0
gunicorn==21.2.0
idna==3.4
kombu==5.3.2
openpyxl==3.1.2
packaging==23.2
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pyowm==3.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sorl-thumbnail==12.9.0
sorl-thumbnail-serializer-field==0.2.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.8
webencodings==0.5.1
XlsxWriter==3.1.2
